# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Traspac Competition",
        "title": "Traspac Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/budget": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve detailed budget information for authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget Management"
                ],
                "summary": "Get user budget details",
                "responses": {
                    "200": {
                        "description": "Budget details retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.Response"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Budget not found",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update existing budget for authenticated user with default (50-30-20) or custom allocation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget Management"
                ],
                "summary": "Update user budget",
                "parameters": [
                    {
                        "description": "Budget data (custom percentages optional)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_request.BudgetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Budget updated successfully",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Budget not found",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create budget for authenticated user with default (50-30-20) or custom allocation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget Management"
                ],
                "summary": "Create user budget",
                "parameters": [
                    {
                        "description": "Budget data (custom percentages optional)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_request.BudgetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Budget created successfully",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Budget already exists",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete the budget for authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget Management"
                ],
                "summary": "Delete user budget",
                "responses": {
                    "200": {
                        "description": "Budget deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.Response"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Budget not found",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/budget/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve detailed budget summary including allocation, usage, and remaining amounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget Management"
                ],
                "summary": "Get detailed budget summary with usage",
                "responses": {
                    "200": {
                        "description": "Budget summary retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.Response"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Budget not found",
                        "schema": {
                            "$ref": "#/definitions/gin-backend-app_internal_dto_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin-backend-app_internal_dto_common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Detailed error message"
                },
                "message": {
                    "type": "string",
                    "example": "Something went wrong"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "gin-backend-app_internal_dto_common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "gin-backend-app_internal_dto_request.BudgetRequest": {
            "type": "object",
            "required": [
                "weekly_income"
            ],
            "properties": {
                "needs_percent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 20
                },
                "savings_percent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 50
                },
                "wants_percent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 30
                },
                "weekly_income": {
                    "type": "number",
                    "example": 1000000
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}